move types; guild handler; www hookup
