--- pkg/api/handlers/item/v1/util.go.txt
@@ -1,1 +1,0 @@
-package v1
