+++ pkg/apis/guild/v1/types.go
@@ -1,0 +1,11 @@
+package v0alpha1
+
+import (
+	"time"
+)
+
+type Guild struct {
+	Id        string
+	Name      string
+	CreatedAt time.Time
+}
